+++ src/widget/checkbox.rs
@@ -25,13 +25,13 @@ impl StyleSheet {
     ///   - a function that will be called when the [`iced::Checkbox`] is toggled. It
     ///     will receive the new state of the [`iced::Checkbox`] and must produce a
     ///     `Message`.
-    pub fn new<Message>(
+    pub fn new<Message, Renderer: iced_native::text::Renderer>(
         &self,
         is_checked: bool,
         label: impl Into<String>,
         f: impl Fn(bool) -> Message + 'static,
-    ) -> iced::Checkbox<'static, Message> {
-        let mut this = iced::Checkbox::new(is_checked, label, f);
+    ) -> iced_native::widget::Checkbox<'static, Message, Renderer> {
+        let mut this = iced_native::widget::Checkbox::new(is_checked, label, f);
         if let Some(width) = self.width {
             this = this.width(width.into());
         }
@@ -53,14 +53,14 @@ impl StyleSheet {
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct Style {
-    active: iced::checkbox::Style,
-    active_checked: iced::checkbox::Style,
-    hovered: iced::checkbox::Style,
-    hovered_checked: iced::checkbox::Style,
+    active: iced_style::checkbox::Style,
+    active_checked: iced_style::checkbox::Style,
+    hovered: iced_style::checkbox::Style,
+    hovered_checked: iced_style::checkbox::Style,
 }
 
-impl iced::checkbox::StyleSheet for Style {
-    fn active(&self, is_checked: bool) -> iced::checkbox::Style {
+impl iced_style::checkbox::StyleSheet for Style {
+    fn active(&self, is_checked: bool) -> iced_style::checkbox::Style {
         if is_checked {
             self.active_checked
         } else {
@@ -68,7 +68,7 @@ impl iced::checkbox::StyleSheet for Style {
         }
     }
 
-    fn hovered(&self, is_checked: bool) -> iced::checkbox::Style {
+    fn hovered(&self, is_checked: bool) -> iced_style::checkbox::Style {
         if is_checked {
             self.hovered_checked
         } else {
@@ -110,7 +110,7 @@ mod de {
     }
 
     impl Inner2 {
-        fn overwrite(&self, style: &mut iced::checkbox::Style) {
+        fn overwrite(&self, style: &mut iced_style::checkbox::Style) {
             if let Some(background) = self.background {
                 style.background = background.into();
             }
@@ -139,24 +139,30 @@ mod de {
         {
             let input = Style::deserialize(deserializer)?;
 
-            let mut active = iced::checkbox::Style {
-                background: iced::Background::Color(iced::Color::from_rgb(0.95, 0.95, 0.95)),
-                checkmark_color: iced::Color::from_rgb(0.3, 0.3, 0.3),
+            let mut active = iced_style::checkbox::Style {
+                background: iced_native::Background::Color(iced_native::Color::from_rgb(
+                    0.95, 0.95, 0.95,
+                )),
+                checkmark_color: iced_native::Color::from_rgb(0.3, 0.3, 0.3),
                 border_radius: 5.0,
                 border_width: 1.0,
-                border_color: iced::Color::from_rgb(0.6, 0.6, 0.6),
+                border_color: iced_native::Color::from_rgb(0.6, 0.6, 0.6),
                 text_color: None,
             };
             input.active.default.overwrite(&mut active);
             let mut active_checked = active;
             input.active.checked.overwrite(&mut active_checked);
 
-            let mut hovered = iced::checkbox::Style {
-                background: iced::Background::Color(iced::Color::from_rgb(0.90, 0.90, 0.90)),
+            let mut hovered = iced_style::checkbox::Style {
+                background: iced_native::Background::Color(iced_native::Color::from_rgb(
+                    0.90, 0.90, 0.90,
+                )),
                 ..active
             };
-            let mut hovered_checked = iced::checkbox::Style {
-                background: iced::Background::Color(iced::Color::from_rgb(0.90, 0.90, 0.90)),
+            let mut hovered_checked = iced_style::checkbox::Style {
+                background: iced_native::Background::Color(iced_native::Color::from_rgb(
+                    0.90, 0.90, 0.90,
+                )),
                 ..active_checked
             };
             input.hovered.default.overwrite(&mut hovered);