+++ src/widget/container.rs
@@ -21,11 +21,11 @@ pub struct StyleSheet {
 
 impl StyleSheet {
     /// Creates an empty [`iced::Container`].
-    pub fn new<'a, Message>(
+    pub fn new<'a, Message, Renderer: iced_native::Renderer>(
         &self,
-        content: impl Into<iced::Element<'a, Message>>,
-    ) -> iced::Container<'a, Message> {
-        let mut this = iced::Container::new(content);
+        content: impl Into<iced_native::Element<'a, Message, Renderer>>,
+    ) -> iced_native::widget::Container<'a, Message, Renderer> {
+        let mut this = iced_native::widget::Container::new(content);
         if let Some(padding) = self.padding {
             this = this.padding(padding);
         }
@@ -55,10 +55,10 @@ impl StyleSheet {
 }
 
 #[derive(Debug, Clone, Copy)]
-pub(crate) struct Style(iced::container::Style);
+pub(crate) struct Style(iced_style::container::Style);
 
-impl iced::container::StyleSheet for Style {
-    fn style(&self) -> iced::container::Style {
+impl iced_style::container::StyleSheet for Style {
+    fn style(&self) -> iced_style::container::Style {
         self.0
     }
 }
@@ -82,7 +82,7 @@ mod de {
             D: Deserializer<'de>,
         {
             let input = Style::deserialize(deserializer)?;
-            let mut this = iced::container::Style::default();
+            let mut this = iced_style::container::Style::default();
 
             if let Some(text_color) = input.text_color {
                 this.text_color = Some(text_color.into());